pyproject.toml
[tool.poetry]
name = "ravager"
version = "1.0.0"
description = "A bot that sails the high seas and brings treasures from the land unknown"
authors = ["Aseem Manna <coolfool149@gmail.com>"]
exclude = ["ravager/.*","ravager/*snippets.py","ravager/services/mega"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.10"
python-dotenv = "^0.19.2"
requests = "^2.27.0"
heroku3 = "^5.1.4"
pytz = "^2021.3"
pexpect = "^4.8.0"
psutil = "^5.9.0"
SQLAlchemy = "^1.4.29"
aria2p = "^0.11.1"
celery = {extras = ["redis"], version = "^5.2.3"}
setuptools = "59.6.0"
Flask = "^2.0.2"
google-api-python-client = "^2.36.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
python-dateutil = "^2.8.2"
torrentool = "^1.1.1"
dataclasses-json = "^0.5.6"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
